address-validator/requirements.txt
Flask
flask_marshmallow
requests
